breaking: add base directory to LsColors.styleForPath
Previously, it was implied that the path provided was relative to
`std.fs.cwd()`. Now, this behavior is explicit and overridable,
allowing for more specific use-cases.

The background for this change is an overhaul of
`EntryType.fromPath`. This function now only calls a `stat` syscall
instead of `open` and `stat`, which was the previous behavior. This
fixes an issue that occurred previously when getting the style for a
broken symlink.